Adds go implementation of deduping cvs by a column
Reads cvs, checks for email in column, sets the column index
wherever 'email' is found. Then parses the cvs in case
it has weird line breaks. Then reads file into a map with
the email address as the key, as maps only take unique 
entires. Stores the value of the map as the line where 
the email was.  Then writes the map to an output file.